requirements.txt
asgiref==3.5.2
attrs==22.1.0
decorator==5.1.1
Django==3.2.16
django-cleanup==6.0.0
django-tinymce==3.5.0
document==1.0
item==0.0.3
mccabe==0.7.0
named==1.1.0
Pillow==9.3.0
pycodestyle==2.9.1
python-dotenv==0.21.0
pytils==0.4.1
pytz==2022.4
selection==0.0.14
self==2020.12.3
six==1.16.0
slug==2.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
typing_extensions==4.4.0
user-agent==0.1.10
weblib==0.1.30
wraps==0.2.0
